turn moduleId into a Long.

+++ java/gadgets/src/main/java/org/apache/shindig/gadgets/GadgetContext.java
@@ -50,7 +50,7 @@ public class GadgetContext {
   /**
    * @return The module id for this request.
    */
-  public int getModuleId() {
+  public long getModuleId() {
     return 0;
   }
